+++ bistro/orders/views.py
@@ -4,7 +4,6 @@ from decimal import Decimal
 from django.contrib import messages
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import F
-from django.db.models import Q
 from django.db.models import Sum
 from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect
@@ -18,6 +17,7 @@ from .forms import TotalPaidOrdersForm
 from .forms import UpdateStatusForm
 from .handlers.add_order import handle_add_order
 from .handlers.delete_order import handle_delete_order
+from .handlers.search_order import handle_search_order
 from .models import Order
 
 
@@ -47,37 +47,6 @@ def order_management_view(request):
         "modals": get_modals(),
     }
     return render(request, "orders/order_management.html", context)
-
-
-def handle_search_order(request):
-    form = SearchOrderForm(request.POST)
-    if form.is_valid():
-        table_number = form.cleaned_data.get("table_number")
-        status = form.cleaned_data.get("status")
-        search_date = form.cleaned_data.get("date")
-
-        search_date_start = timezone.make_aware(
-            datetime.combine(search_date, datetime.min.time()),
-        )
-        search_date_end = timezone.make_aware(
-            datetime.combine(search_date, datetime.max.time()),
-        )
-
-        filters = Q(created_at__range=[search_date_start, search_date_end])
-
-        if table_number:
-            filters &= Q(table_number=table_number)
-
-        if status != "all":
-            filters &= Q(status=status)
-
-        orders = Order.objects.filter(filters).annotate(
-            total_order_price=Sum(
-                F("order_items__menu_item__price") * F("order_items__quantity"),
-            ),
-        )
-        return orders, True, True
-    return Order.objects.all(), False, False
 
 
 def handle_update_status(request):
